+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 - TYPO3 helhum console
+- TSFE subscriber to load TYPO3 stuff for Symfony routes
 
 ## [5.1.0] - 2024-10-04
 ### Added